# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finance App API",
        "description": "API para organização financeira",
        "contact": {
            "email": "contato@financeapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "List all transactions of the authenticated user",
                "operationId": "9b50e29e364179ebde24ac5fc59cd442",
                "responses": {
                    "200": {
                        "description": "List of transactions"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "operationId": "54eaa3952808f639d46f969f4c8b98c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "amount",
                                    "date",
                                    "type"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Compra no mercado"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-06"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "debito",
                                        "enum": [
                                            "credito",
                                            "debito"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pago",
                                        "enum": [
                                            "pago",
                                            "pendente"
                                        ]
                                    },
                                    "observations": {
                                        "type": "string",
                                        "example": "Compra de itens essenciais"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction details",
                "operationId": "45e35076645e40d84080bf253c06a689",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update a transaction",
                "operationId": "43fb5702f8299f2d383e663429920d7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Compra no mercado"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-06"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "debito",
                                        "enum": [
                                            "credito",
                                            "debito"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pago",
                                        "enum": [
                                            "pago",
                                            "pendente"
                                        ]
                                    },
                                    "observations": {
                                        "type": "string",
                                        "example": "Compra de itens essenciais"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction updated"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Delete a transaction",
                "operationId": "bc63b06ba9e12cac8ef887c638602dd5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}